refactor: move lms utilities to separate file
